requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asgiref==3.11.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.4
callbacks==0.3.0
camelot-py==1.0.9
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.7.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.12
django-googlesearch==0.2
docstring_parser==0.17.0
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.25.2
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googlesearch-python==1.3.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
idna==3.11
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-openai==1.1.10
langchain-postgres==0.0.17
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.12
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.17.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.3
numpy==2.2.6
openai==2.24.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
param==2.3.2
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.3.6
pillow==12.1.1
plotly==6.6.0
propcache==0.4.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
px==0.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==3.17.4
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyviz_comms==3.0.6
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
retrievers==0.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
streamlit==1.55.0
streamlit-chat==0.1.1
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==1.13.1
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
types-pytz==2026.1.1.20260304
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.32
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
